feat: status 400 error res added to ID related reqs

+++ src/app/service/ProductService.ts
@@ -1,5 +1,8 @@
+import BadRequestError from '../errors/BadRequestError'
+import { Types } from 'mongoose'
 import { IProductResponse, IProduct } from '../interfaces/IProduct'
 import ProductRepository from '../repository/ProductRepository'
+
 
 class ProductService {
   async create (payload: IProduct): Promise<IProductResponse> {
@@ -8,6 +11,8 @@ class ProductService {
   }
 
   async update (payload: IProduct, productId: string): Promise<IProductResponse | null> {
+    if (!Types.ObjectId.isValid(productId)) throw new BadRequestError('Not an valid ID');
+
     payload.stock_control_enabled = payload.qtd_stock > 0
     payload.updatedAt = new Date()
 
@@ -16,6 +21,8 @@ class ProductService {
   }
 
   async patch (payload: IProduct, productId: string): Promise<IProductResponse | null> {
+    if (!Types.ObjectId.isValid(productId)) throw new BadRequestError('Not an valid ID');
+    
     payload.stock_control_enabled = payload.qtd_stock > 0
     payload.updatedAt = new Date()
 
@@ -34,11 +41,15 @@ class ProductService {
   }
 
   async getOne (productId: string) {
+    if (!Types.ObjectId.isValid(productId)) throw new BadRequestError('Not an valid ID');
+    
     const result = await ProductRepository.getOne(productId)
     return result
   }
 
   async delete (productId: string) {
+    if (!Types.ObjectId.isValid(productId)) throw new BadRequestError('Not an valid ID');
+
     const result = await ProductRepository.delete(productId)
     return result
   }
